usb_host: Move to idf-extra-components
