+++ usb/test_app/CMakeLists.txt
@@ -1,0 +1,11 @@
+# The following lines of boilerplate have to be in your project's
+# CMakeLists in this exact order for cmake to work correctly
+cmake_minimum_required(VERSION 3.16)
+
+set(EXTRA_COMPONENT_DIRS $ENV{IDF_PATH}/examples/peripherals/usb/host/cdc/common
+                         $ENV{IDF_PATH}/examples/peripherals/usb/host/msc/components/)
+
+# Set the components to include the tests for.
+set(TEST_COMPONENTS "cdc_acm_host" "msc" CACHE STRING "List of components to test")
+include($ENV{IDF_PATH}/tools/cmake/project.cmake)
+project(usb_test_app)
