+++ usb/test_app/main/CMakeLists.txt
@@ -1,0 +1,3 @@
+idf_component_register(SRCS "usb_test_main.c"
+                       INCLUDE_DIRS ""
+                       REQUIRES unity driver usb)
