+++ usb/usb_host_cdc_acm/idf_component.yml
@@ -1,0 +1,8 @@
+version: "1.0.0"
+description: USB Host CDC-ACM driver
+url: https://github.com/espressif/idf-extra-components/tree/master/usb/usb_host_cdc_acm
+dependencies:
+  idf: ">=4.4"
+targets:
+  - esp32s2
+  - esp32s3
