+++ usb/usb_host_cdc_acm/include/usb/usb_types_cdc.h
